prototype/requirements.txt
# CRM Digital FTE - Prototype

requirements.txt==4.0.0
python-dotenv==1.0.0
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
openai==1.3.0
tiktoken==0.5.2
chromadb==0.4.18
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
aiohttp==3.9.1
twilio==8.10.0
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.111.0
kafka-python==2.0.2
textblob==0.17.1
nltk==3.8.1
